+++ srcs/MineClearance.Infrastructure/Constants.cs
@@ -1,3 +1,4 @@
+using MineClearance.Infrastructure.Models;
 using System;
 using System.IO;
 using System.Text.Json;
@@ -20,6 +21,13 @@ public static class Constants
     };
 
     /// <summary>
+    /// 要使用的路径比较器
+    /// </summary>
+    public static readonly StringComparer PathComparer = OperatingSystem.IsWindows()
+        ? StringComparer.OrdinalIgnoreCase
+        : StringComparer.Ordinal;
+
+    /// <summary>
     /// 程序数据根目录
     /// </summary>
     public static readonly string AppDataRootDirectory = Path.Combine(
@@ -28,57 +36,108 @@ public static class Constants
     );
 
     /// <summary>
-    /// 数据目录文件夹
+    /// 数据目录文件夹名
     /// </summary>
-    public const string DataDirectory = "Data";
+    public const string DataDirectoryName = "Data";
 
     /// <summary>
-    /// 日志文件夹
+    /// 日志文件夹名
     /// </summary>
-    public const string LogDirectory = "Logs";
+    public const string LogDirectoryName = "Logs";
 
     /// <summary>
-    /// 设置文件夹
+    /// 设置文件夹名
     /// </summary>
-    public const string SettingsDirectory = "Settings";
+    public const string SettingsDirectoryName = "Settings";
+
+    /// <summary>
+    /// 更新数据目录
+    /// </summary>
+    public static readonly string UpdateDataDirectory = Path.Combine(AppDataRootDirectory, "Update");
 
     /// <summary>
     /// 游戏存档文件路径
     /// </summary>
     public static readonly string GameSaveDataFilePath = Path.Combine(
-        AppDataRootDirectory, DataDirectory, "SavedGame.json"
+        AppDataRootDirectory, DataDirectoryName, $"SavedGame{JsonFileSuffix}"
     );
 
     /// <summary>
     /// 游戏结果记录文件路径
     /// </summary>
     public static readonly string GameResultsFilePath = Path.Combine(
-        AppDataRootDirectory, DataDirectory, "History.json"
+        AppDataRootDirectory, DataDirectoryName, $"History{JsonFileSuffix}"
     );
 
     /// <summary>
     /// 最新日志文件路径
     /// </summary>
     public static readonly string LatestLogFilePath = Path.Combine(
-        AppDataRootDirectory, LogDirectory, $"Latest{LogFileSuffix}"
+        AppDataRootDirectory, LogDirectoryName, $"Latest{LogFileSuffix}"
     );
 
     /// <summary>
     /// 日志设置文件路径
     /// </summary>
     public static readonly string LogSettingsFilePath = Path.Combine(
-        AppDataRootDirectory, SettingsDirectory, $"LogSettings{SettingFileSuffix}"
+        AppDataRootDirectory, SettingsDirectoryName, $"LogSettings{JsonFileSuffix}"
     );
 
     /// <summary>
-    /// 设置文件后缀
+    /// 备份目录
     /// </summary>
-    public const string SettingFileSuffix = ".json";
+    public static readonly string BackupDirectory = Path.Combine(UpdateDataDirectory, "Backup");
+
+    /// <summary>
+    /// 引导副本所在的目录
+    /// </summary>
+    public static readonly string BootstrapCopyDirectory = Path.Combine(UpdateDataDirectory, "BootstrapCopy");
+
+    /// <summary>
+    /// 下载的更新包的文件路径
+    /// </summary>
+    public static readonly string UpdatePackageFilePath = Path.Combine(
+        UpdateDataDirectory, $"Update{ZipFileSuffix}"
+    );
+
+    /// <summary>
+    /// 新版本的版本号文件路径
+    /// </summary>
+    public static readonly string NewVersionFilePath = Path.Combine(UpdateDataDirectory, "NewVersion.txt");
+
+    /// <summary>
+    /// 更新信息文件路径
+    /// </summary>
+    public static readonly string UpdateInfoFilePath = Path.Combine(
+        UpdateDataDirectory, $"{nameof(UpdateInfo)}{JsonFileSuffix}"
+    );
+
+    /// <summary>
+    /// 更新日志文件路径
+    /// </summary>
+    public static readonly string UpdateLogFilePath = Path.Combine(
+        UpdateDataDirectory, $"Update{LogFileSuffix}"
+    );
 
     /// <summary>
     /// 日志文件后缀
     /// </summary>
     public const string LogFileSuffix = ".log";
+
+    /// <summary>
+    /// json 文件后缀
+    /// </summary>
+    public const string JsonFileSuffix = ".json";
+
+    /// <summary>
+    /// zip 文件后缀
+    /// </summary>
+    public const string ZipFileSuffix = ".zip";
+
+    /// <summary>
+    /// 使用引导更新模式的参数
+    /// </summary>
+    public const string UseBootstrapUpdateModeArgument = "--use-bootstrap-update-mode";
 
     /// <summary>
     /// 最大日志文件数量